Move javarosawrapper
