+++ collect_app/src/main/java/org/odk/collect/android/formentry/FormControllerProvider.java
@@ -1,6 +1,6 @@
 package org.odk.collect.android.formentry;
 
-import org.odk.collect.android.formentry.javarosawrapper.FormController;
+import org.odk.collect.android.javarosawrapper.FormController;
 
 public interface FormControllerProvider {
