+++ collect_app/src/main/java/org/odk/collect/android/formentry/saving/FormSaver.java
@@ -3,7 +3,7 @@ package org.odk.collect.android.formentry.saving;
 import android.net.Uri;
 
 import org.odk.collect.android.analytics.Analytics;
-import org.odk.collect.android.formentry.javarosawrapper.FormController;
+import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.tasks.SaveToDiskResult;
 
 public interface FormSaver {
