+++ collect_app/src/main/java/org/odk/collect/android/upload/InstanceGoogleSheetsUploader.java
@@ -62,7 +62,7 @@ import java.util.regex.Pattern;
 
 import timber.log.Timber;
 
-import static org.odk.collect.android.formentry.javarosawrapper.FormController.INSTANCE_ID;
+import static org.odk.collect.android.javarosawrapper.FormController.INSTANCE_ID;
 
 public class InstanceGoogleSheetsUploader extends InstanceUploader {
     private static final String PARENT_KEY = "PARENT_KEY";
