+++ collect_app/src/test/java/org/odk/collect/android/javarosawrapper/FormControllerTest.java
@@ -1,4 +1,4 @@
-package org.odk.collect.android.formentry.javarosawrapper;
+package org.odk.collect.android.javarosawrapper;
 
 import com.google.common.io.Files;
