+++ collect_app/src/test/java/org/odk/collect/android/widgets/OSMWidgetTest.java
@@ -18,7 +18,7 @@ import org.junit.Test;
 import org.mockito.Mock;
 import org.odk.collect.android.R;
 import org.odk.collect.android.formentry.questions.QuestionDetails;
-import org.odk.collect.android.formentry.javarosawrapper.FormController;
+import org.odk.collect.android.javarosawrapper.FormController;
 import org.odk.collect.android.widgets.base.BinaryWidgetTest;
 
 import java.io.File;
